Big overhaul of the projects module. Factors hard coded paths and IDs into a configuration object under the C1V0 namespace. Also factors out most of the remaining scripts floating around the global namespace. This is work toward #15.
