+++ README.md
@@ -6,7 +6,9 @@
   <img src="app/images/h-logo.png" alt="www.chrisvogt.me">
 </p>
 
-My personal website and project portfolio, based on a [Moqups design](https://app.moqups.com/chris@artinreality.com/81jSoAGytP/view/page/add529438) I created first, though it has gone through many iterations since. Stats data is fetched from [stats.chrisvogt.me](https://stats.chrisvogt.me) and the project data is in [Firebase](https://firebase.google.com/).
+My personal website and project portfolio, originally based on simple [Moqups design](https://app.moqups.com/chris@artinreality.com/81jSoAGytP/view/page/add529438), which it has gone through _many_ iterations.
+
+Stats data is fetched from an API behind [stats.chrisvogt.me](https://stats.chrisvogt.me). The project data is stored in a [Firebase](https://firebase.google.com/) NoSQL database.
 
 _View it live at [www.chrisvogt.me](https://www.chrisvogt.me)._
 
@@ -34,6 +36,6 @@ _View it live at [www.chrisvogt.me](https://www.chrisvogt.me)._
   <img src="https://cdn.rawgit.com/gulpjs/artwork/master/gulp-2x.png" alt="Gulp" height="48">
   <img src="https://avatars0.githubusercontent.com/u/1335026?v=3&s=200" alt="Firebase" height="48">
 	<img src="http://upload.wikimedia.org/wikipedia/en/9/9e/JQuery_logo.svg" alt="jQuery" height="48">
-	<img src="https://upload.wikimedia.org/wikipedia/commons/1/1b/CSS3_and_HTML5_badges.svg" alt="HTML5 &amp; CSS3" height="48">
+	<img src="https://upload.wikimedia.org/wikipedia/commons/1/10/CSS3_and_HTML5_logos_and_wordmarks.svg" alt="HTML5 &amp; CSS3" height="48">
 	<img src="https://cdn.rawgit.com/mathamoz/ionic-builder/898ac76dc9e9edeb02d1825358eca95ec742b985/public/images/why-the-yeti.svg" alt="Zurb Foundation" height="48">
 </p>
