+++ app/index.html
@@ -7,9 +7,9 @@
     www.chrisvogt.me » @c1v0
     https://github.com/chrisvogt/www.chrisvogt.me
  -->
-    <title>Chris Vogt — Front End Developer &amp; engineering student ✌ SF ➳ PHX️</title>
+    <title>Chris Vogt — Front End Developer ✌ SF ➳ PHX️</title>
 
-    <meta name="description" content="Front End Developer and engineering student in Phoenix, sharing open source projects written in JavaScript and PHP.">
+    <meta name="description" content="Front End Developer in Phoenix, sharing open source projects written in JavaScript and PHP.">
     <link rel="author" href="humans.txt">
 
     <meta name="theme-color" content="#4527a0">
@@ -19,15 +19,15 @@
     <link rel="apple-touch-icon" sizes="120x120" href="/images/favicon-120.png">
     <link rel="apple-touch-icon" sizes="152x152" href="/images/favicon-152.png">
 
-    <meta name="og:title" content="Chris Vogt — Front End Developer &amp; engineering student ✌ SF ➳ PHX️">
-    <meta name="og:description" content="Front End Developer and engineering student in Phoenix, sharing open source projects written in JavaScript and PHP.">
+    <meta name="og:title" content="Chris Vogt — Front End Developer ✌ SF ➳ PHX️">
+    <meta name="og:description" content="Front End Developer in Phoenix, sharing open source projects written in JavaScript and PHP.">
     <meta name="og:image" content="https://res.cloudinary.com/chrisvogt/image/upload/v1433903064/og-banner_saphne.jpg"/>
 
     <meta name="twitter:card" content="summary_large_image">
     <meta name="twitter:site" content="@c1v0">
     <meta name="twitter:creator" content="@c1v0">
-    <meta name="twitter:title" content="Chris Vogt — Front End Developer &amp; engineering student ✌ SF ➳ PHXX">
-    <meta name="twitter:description" content="Front End Developer and engineering student in Phoenix, sharing open source projects written in JavaScript and PHP.">
+    <meta name="twitter:title" content="Chris Vogt — Front End Developer ✌ SF ➳ PHXX">
+    <meta name="twitter:description" content="Front End Developer in Phoenix, sharing open source projects written in JavaScript and PHP.">
     <meta name="twitter:image" content="https://res.cloudinary.com/chrisvogt/image/upload/v1433903064/og-banner_saphne.jpg">
 
     <script type="application/ld+json">
@@ -58,7 +58,6 @@
     <!-- bower:css -->
     <link rel="stylesheet" href="/bower_components/remodal/dist/remodal.css">
     <link rel="stylesheet" href="/bower_components/remodal/dist/remodal-default-theme.css">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/foundation/5.5.3/css/foundation.min.css">
     <link rel="stylesheet" href="styles/main.css">
     <!-- endbower -->
     <!-- endbuild -->
@@ -74,7 +73,7 @@
         </ul>
         <section class="top-bar-section">
           <ul class="right">
-            <li><a href="http://stats.chrisvogt.me" title="Coding stats, last 30 days"><i class="fa fa-area-chart"></i> Stats</a></li>
+            <li><a href="https://stats.chrisvogt.me" title="Coding stats, last 30 days"><i class="fa fa-area-chart"></i> Stats</a></li>
             <li><a href="https://resume.chrisvogt.me" title="My résumé on StackOverflow Careers"><i class="fa fa-briefcase"></i> Résumé</a></li>
           </ul>
           <ul class="left">
@@ -221,8 +220,8 @@
             </div>
             <h2>About Me</h2>
 
-            <p>In the real world I inhabit Phoenix, Arizona, USA. During the day, I'm a Front End Developer building healthcare payment processing solutions. Most weekday nights I can be spotted near N33° 29.0744&#96;, W112° 5.3194&#96;.</p>
-            <p>Over the last decade I've been a developer for electronic discovery, healthcare, and education companies. I have been enrolled in night courses at Phoenix College every semester since Fall 2014, and am a member of both the Phi Theta Kappa Honor Society and the IEEE Society on Social Implications of Technology.</p>
+            <p>In the real world I inhabit Phoenix, Arizona, USA. During the day, I'm a Front End Developer building healthcare payment processing solutions. Weekday nights I can often be spotted near N33° 29.0744&#96;, W112° 5.3194&#96;.</p>
+            <p>Over the last decade I've been a developer for electronic discovery, healthcare, and education technology companies working on intranets, web portals, and learning management systems. I have been enrolled in night courses at Phoenix College every semester since Fall 2014, and am a member of both the Phi Theta Kappa Honor Society and the IEEE Society on Social Implications of Technology.</p>
             <p>I'm a multifaceted individual with a strong interest in the web, music, literature, and science. I'm motivated by an innate drive to both understand and create. To learn more, check out my <a href="https://github.com/chrisvogt" title="Chris Vogt on GitHub">code</a> on GitHub, <a href="https://youtu.be/OoT70uJpB5w?list=PLNpEtdDfoma2QoX2Oxn8tvybYoiQI_yz_" title="Chris Vogt piano YouTube channel">piano channel</a> on YouTube, and Sci Phi <a href="https://soundcloud.com/findmebymidnight/sets/starry" title="Find Me By Midnight on SoundCloud">electronic music</a> on SoundCloud.</p>
           </div>
         </div>
@@ -253,13 +252,13 @@
       </div>
     </section>
 
-    <footer>
+    <footer id="docfoot">
       <div class="row">
-        <div class="small-6 columns">
+        <div class="small-12 medium-6 columns">
           <p>Made with <span class="octicon octicon-code"></span> + <span class="octicon octicon-heart"></span> by <a href="https://https://twitter.com/c1v0" title="@C1V0 on Twitter">@C1V0</a> in Phoenix 🌵</p>
         </div>
-        <div class="small-6 columns">
-          <ul class="inline-list right">
+        <div class="small-12 medium-6 columns">
+          <ul class="inline-list">
             <li><a href="http://stats.chrisvogt.me" title="Chris Vogt's open source stats">Stats</a></li>
             <li><a href="https://resume.chrisvogt.me" title="Chris Vogt's résumé">Résumé</a></li>
           </ul>
@@ -309,14 +308,6 @@
     </div>
 
     <script>
-      window.ga=window.ga||function(){(ga.q=ga.q||[]).push(arguments)};ga.l=+new Date;
-      ga('create', 'UA-33558417-1', 'auto', {'allowLinker': true});
-      ga('require', 'linker');
-      ga('linker:autoLink', ['stats.chrisvogt.me']);
-      ga('send', 'pageview');
-    </script>
-
-    <script>
       WebFontConfig = {
         google: { families: [ 'Quicksand::latin', 'Merriweather:300,300italic:latin' ] }
       };
@@ -353,13 +344,13 @@
     <script src="/bower_components/foundation/js/foundation/foundation.tooltip.js"></script>
     <!-- main -->
     <script src="scripts/app.js"></script>
-    <script src="scripts/common.js"></script>
     <script src="scripts/HttpSocket.js"></script>
     <script src="scripts/navigation.jquery.js"></script>
     <script src="scripts/stats.jquery.js"></script>
     <script src="scripts/quotes.jquery.js"></script>
     <script src="scripts/projects.jquery.js"></script>
     <script src="scripts/social.jquery.js"></script>
+    <script src="scripts/analytics.jquery.js"></script>
     <!-- endbower -->
     <!-- endbuild -->
   </body>
