+++ app/styles/base/_typography.scss
@@ -2,36 +2,44 @@
  * Typography
  */
 
-// -- common
+/* common */
 body {
     color: $text-raven;
     font-family: 'Merriweather', serif;
     font-size: 100%;
 }
+
 .text-small {
-  font-size: .875em;
+    font-size: .875em;
 }
 
-// -- accents
+/* accents */
 hr {
     text-align: center;
     margin: 0 auto;
     opacity: 0.5;
 }
+
 ul {
     text-transform: uppercase;
 }
-ul.inline-list.right {
+
+ul.inline-list.right,
+footer ul.inline-list {
     text-transform: none;
 }
-ul.inline-list.right a {
+
+ul.inline-list.right a,
+footer ul.inline-list {
     color: #000;
 }
-ul.inline-list.right a:hover {
+
+ul.inline-list.right a:hover,
+footer ul.inline-list {
     color: rgba(0, 0, 0, 0.5);
 }
 
-// -- headlines
+/* headlines */
 h1 {
     font-family: 'Quicksand', sans-serif;
     font-size: 2.124em;
@@ -39,11 +47,13 @@ h1 {
     line-height: 28px;
     margin-top: 16px;
 }
+
 @media #{$medium-up} {
-  h1 {
-    font-size: 3.124em;
-  }
+    h1 {
+        font-size: 3.124em;
+    }
 }
+
 h2 {
     font-family: "Helvetica Neue", Helvetica, Arial, sans-serif;
     margin-bottom: 1em;
@@ -53,6 +63,7 @@ h2 {
     line-height: 34px;
     font-weight: 400;
 }
+
 h3 {
     font-family: "Helvetica Neue", Helvetica, Arial, sans-serif;
     font-size: 24px;
@@ -62,12 +73,14 @@ h3 {
     font-weight: 500;
     padding-bottom: 10px;
 }
+
 h4 {
     font-family: "Helvetica Neue", Helvetica, Arial, sans-serif;
     color: #000;
     font-size: 28px;
     padding-top: 5px;
 }
+
 blockquote {
     text-decoration: none;
     text-transform: none;
@@ -77,7 +90,8 @@ blockquote {
     font-size: 1.125em;
     line-height: 1.5em;
 }
+
 footer,
-ul.inline-list.right a {
+ul.inline-list a {
     color: #afafaf;
 }
