+++ app/styles/main.scss
@@ -3,6 +3,7 @@ $octicons-font-path: "../fonts/";
 
 // bower:scss
 @import "bower_components/font-awesome/scss/font-awesome.scss";
+@import "bower_components/foundation/scss/foundation.scss";
 @import "bower_components/octicons/octicons/octicons.scss";
 // endbower
